1- Added HomePageView This is a promotional or introductory page for the website.
2- Modified Navbar
It is now dynamic, so its properties can change depending on the type of user.

3- Modified Footer
I hid the current data to modify it later. Some data will appear when the user is on the HomePageView.

4- Added PostsPageView
A page that displays all job posts. Data can be displayed as a grid or a list. Filters will be added in the future; currently, it’s just the layout:

postGrid

PostList
This is part of the page.

5- Added PostPageView
A page to display a specific post.

----
6- Added router + index.js
It will not be used currently.