1- Added ProfilePageView A page that displays the user’s profile. If the user is an admin in a company, the company they work for will also be displayed.
2- Modified App
Added the user’s state and passed it to Navbar and ProfilePageView.

3- Modified PostGrid and PostList
Now images are included to visualize the post layout, with some adjustments to dimensions.

4- Modified index.js
Added ProfilePageView and PostPageView, along with adding the site name in the extension.

5- Modified PostPageView
Added images for the post and the company.

6- Modified PostsPageView
Changed the way PostGrid and PostList are called, from if to show.
This way, the component is called once instead of twice, and the display method is switched without fetching the same data multiple times.