Latest Updates
Added a new page: EditCompanyPageView to edit company information.

Removed some unnecessary comments in EditProfilePageView.

Minor adjustments in ProfilePageView, such as changing "Followed" to "Unfollow".

Passed user permissions to PostProfile and updated the design of PostProfile and Projects.

Additions and adjustments in CompanyPageView.

Modified the display of employees and the company manager, and updated the design of PostProfile and Projects.

Added a new section for Required Jobs.