link4job
data.json
Improve the data structure.

CommentCard
Improve the way data is fetched and how comments are displayed.

FriendCard / UserSearchCard
Unify the display style of user profile images.

JobsGridCard / JobsListCard
Fix display issues related to images and sizing.

PostsCard
Rewrite the data-fetching and display logic to be simpler and less complex,
and fix the previous issue where comments had to be rendered as part of the card itself.

ProfilePageView
Add a button to navigate to the company page.

CompanyPageView
Improve the display of the account banner.

Navbar
Fix issues related to the adminAccount user type.